third_party/blink/web_tests/external/wpt/dom/parts/basic-dom-part-declarative-brace-syntax.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] Just parts, no text before
  assert_equals: expected "<!----><!---->" but got " <!----> <!---->"
[FAIL] Foreign content should support Parts
  assert_equals: expected 1 but got 0
Harness: the test ran to completion.
